+++ documentation/version-history.md
@@ -6,4 +6,5 @@ version history                 {#version_history}
  0.1.0          |                                                      | initial release
  0.1.1          | [4](https://github.com/ZEISS/czicheck/pull/4)        | update of metadata-schema (compression)
  0.1.2          | [9](https://github.com/ZEISS/czicheck/pull/9)        | add checker "subblkbitmapvalid"
- 0.1.3          | [11](https://github.com/ZEISS/czicheck/pull/11)      | disallow DTD-processing (prevent XXE-vulnerability)
+ 0.2.0          | [10](https://github.com/ZEISS/czicheck/pull/10)      | add checker "topographymetadata"
+ 0.2.1          | [11](https://github.com/ZEISS/czicheck/pull/11)      | disallow DTD-processing (prevent XXE-vulnerability)